+++ source/win-main/store.ts
@@ -19,8 +19,8 @@ import sanitizeHtml from 'sanitize-html'
 import md2html from '@common/util/md-to-html'
 import sort from '../main/modules/fsal/util/sort'
 import { CodeFileMeta, DirMeta, MDFileMeta } from '@dts/common/fsal'
+import path from 'path'
 
-const path = (window as any).path
 const ipcRenderer = (window as any).ipc as Electron.IpcRenderer
 
 interface FSALEvent {