+++ source/win-project-properties/App.vue
@@ -95,6 +95,7 @@ import TextControl from '@common/vue/form/elements/Text.vue'
 import { IpcRenderer } from 'electron'
 import { defineComponent } from 'vue'
 import { ProjectSettings } from '@dts/common/fsal'
+import { WindowTab } from '@dts/renderer/window'
 
 const ipcRenderer: IpcRenderer = (window as any).ipc
 
@@ -120,19 +121,17 @@ export default defineComponent({
       tabs: [
         {
           id: 'formats-control',
-          target: 'formats-panel',
           label: 'General',
           icon: 'cog',
           controls: 'formats-panel'
         },
         {
           id: 'files-control',
-          target: 'files-panel',
           label: 'Files',
           icon: 'file-settings',
           controls: 'formats-panel'
         }
-      ],
+      ] as WindowTab[],
       currentTab: 0
     }
   },